added mount point test to os_utils.sh

+++ util/os_utils.sh
@@ -6,11 +6,18 @@ pst_debug_echo "$BASH_SOURCE"
 # unique to OS
 case "$PST_OS" in
 	"darwin")
-		if [ -n "$(command -v defaults 2>/dev/null)" ] ; then
+		
+		if [ -n "$(command -v mount 2>/dev/null)" ] && [ -n "$(command -v grep 2>/dev/null)" ] ; then
+			function testmount() {
+				[ -d $1 ] && mount | grep -qs "on $1" > /dev/null
+			}
+		fi
+				
+		if [ -n "$(command -v defaults 2>/dev/null)" ] && [ -n "$(command -v killall 2>/dev/null)" ] ; then
 			alias hide='defaults write com.apple.finder AppleShowAllFiles FALSE && killall Finder'
 			alias unhide='defaults write com.apple.finder AppleShowAllFiles TRUE && killall Finder'
 		fi
-
+		
 		# I could add to the existing OS the components called out below:
 			# lesspipe
 			# bash_completion
@@ -21,7 +28,15 @@ case "$PST_OS" in
 	
 	
 	"linux")
-		# make less more friendly for non-text input files, see lesspipe(1)
+				
+		if [ -f /proc/mounts ] && [ -n "$(command -v grep 2>/dev/null)" ] ; then
+			function testmount() {
+				[ -d $1 ] && grep -qs "$1" /proc/mounts > /dev/null
+			}
+		fi
+		
+		
+						# make less more friendly for non-text input files, see lesspipe(1)
 		[ -x /usr/bin/lesspipe ] && eval "$(SHELL=/bin/sh lesspipe)"
 	
 		# enable programmable completion features (you don't need to enable
